Introduce HoverablePopover
