+++ components/hoverable-popover.module.css
@@ -1,0 +1,8 @@
+.hoverablePopover {
+    border: 1px solid var(--theme-toolbarActive)
+}
+
+.hoverablePopBody {
+    font-weight: 500;
+    font-size: 0.9rem;
+}
