+++ components/user-popover.js
@@ -1,14 +1,11 @@
 import { USER } from '@/fragments/users'
 import errorStyles from '@/styles/error.module.css'
 import { useLazyQuery } from '@apollo/client'
-import Link from 'next/link'
-import { useRef, useState } from 'react'
-import { Popover } from 'react-bootstrap'
-import OverlayTrigger from 'react-bootstrap/OverlayTrigger'
-import { UserBase, UserSkeleton } from './user-list'
-import styles from './user-popover.module.css'
 import classNames from 'classnames'
+import Link from 'next/link'
+import HoverablePopover from './hoverable-popover'
 import ItemPopover from './item-popover'
+import { UserBase, UserSkeleton } from './user-list'
 
 function StackingSince ({ since }) {
   return (
@@ -26,8 +23,6 @@ function StackingSince ({ since }) {
 }
 
 export default function UserPopover ({ name, children }) {
-  const [showOverlay, setShowOverlay] = useState(false)
-
   const [getUser, { loading, data }] = useLazyQuery(
     USER,
     {
@@ -36,52 +31,19 @@ export default function UserPopover ({ name, children }) {
     }
   )
 
-  const timeoutId = useRef(null)
-
-  const handleMouseEnter = () => {
-    clearTimeout(timeoutId.current)
-    getUser()
-    timeoutId.current = setTimeout(() => {
-      setShowOverlay(true)
-    }, 500)
-  }
-
-  const handleMouseLeave = () => {
-    clearTimeout(timeoutId.current)
-    timeoutId.current = setTimeout(() => setShowOverlay(false), 100)
-  }
-
   return (
-    <OverlayTrigger
-      show={showOverlay}
-      placement='bottom'
-      onHide={handleMouseLeave}
-      overlay={
-        <Popover
-          onMouseEnter={handleMouseEnter}
-          onMouseLeave={handleMouseLeave}
-          className={styles.userPopover}
-        >
-          <Popover.Body className={styles.userPopBody}>
-            {!data || loading
-              ? <UserSkeleton />
-              : !data.user
-                  ? <h1 className={classNames(errorStyles.status, errorStyles.describe)}>USER NOT FOUND</h1>
-                  : (
-                    <UserBase user={data.user} className='mb-0 pb-0'>
-                      <StackingSince since={data.user.since} />
-                    </UserBase>
-                    )}
-          </Popover.Body>
-        </Popover>
-      }
-    >
-      <span
-        onMouseEnter={handleMouseEnter}
-        onMouseLeave={handleMouseLeave}
-      >
-        {children}
-      </span>
-    </OverlayTrigger>
+    <HoverablePopover
+      onShow={getUser}
+      trigger={children}
+      body={!data || loading
+        ? <UserSkeleton />
+        : !data.user
+            ? <h1 className={classNames(errorStyles.status, errorStyles.describe)}>USER NOT FOUND</h1>
+            : (
+              <UserBase user={data.user} className='mb-0 pb-0'>
+                <StackingSince since={data.user.since} />
+              </UserBase>
+              )}
+    />
   )
 }
